+++ packages/frontend-shared/cypress/e2e/support/e2eProjectDirs.ts
@@ -114,6 +114,7 @@ export const e2eProjectDirs = [
   'unify-onboarding-with-config',
   'unify-plugin-errors',
   'various-file-types',
+  'vite-ct',
   'webpack-preprocessor',
   'webpack-preprocessor-awesome-typescript-loader',
   'webpack-preprocessor-ts-loader',