+++ packages/launchpad/src/setup/Wizard.vue
@@ -28,13 +28,13 @@ import { computed, ref } from 'vue'
 import LaunchpadHeader from './LaunchpadHeader.vue'
 import { useI18n } from '@cy/i18n'
 
+export type CurrentStep = 'selectFramework' | 'installDependencies'
+
 const props = defineProps<{
   gql: WizardFragment
 }>()
 
 const { t } = useI18n()
-
-export type CurrentStep = 'selectFramework' | 'installDependencies'
 
 const currentStep = ref<CurrentStep>('selectFramework')